refactor(ledger): drop Huma suffixes on organization :hammer:
The six exported fiber.Ctx terminals on OrganizationHandler were unreachable —
registerOrganizationRoutesToApp passes middleware only and hands the terminals
to Huma.

- Delete the dead Fiber wrappers and their tests.
- Port the branches those tests covered onto the Huma transport; all six
  organization.go cores now measure 100%, up from a Huma-only run that left
  updateOrganization at 0% and getAllOrganizations at 62.5%.
  getAllOrganizations now exceeds its 90.6% baseline: the metadata/name-filter
  mutual-exclusion guard was uncovered on both transports before.
- Rename methods and envelope types: CreateOrganizationHuma ->
  CreateOrganization, ListOrganizationsInputHuma -> ListOrganizationsRequest,
  and so on.
- organization_handler_huma.go -> organization_handler.go,
  organization_huma_test.go -> organization_handler_test.go.

permissions.yaml annotates each permission block with the file:line that
registers it. This updates organization's citation and the four the earlier
renames in this branch left dangling (account, ledger, portfolio, segment).
The accounttype, asset-rate, metadata and count citations still point at
*_handler_huma.go files that have not been converted yet.

No route, operation ID, or wire contract changes: check-docs reports no
drift.

X-Lerian-Ref: 0x1
